hash/salt the user `sub` field used in KV/cookies
using the webcrypto api, we can store a hashed/salted version of the `sub` user id using the web crypto api. this commit updates `auth0.js` to use this, and updates `readme.md` with the necessary documentation for the new `SALT` secret.

thanks to @dhaynespls for help on implementing this!